refactor: remove unused utility modules and consolidate functionality; add new utilities for version validation and release management
