--- bin/constants/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  ...require('./colors'),
-}
