--- bin/utils/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  ...require('./exec'),
-  ...require('./path'),
-  ...require('./question'),
-}
