+++ src/shared/lib/electron-app/release/utils/extractors.ts
@@ -1,4 +1,4 @@
-exports.extractOwnerAndRepoFromGitRemoteURL = (url) => {
+export function extractOwnerAndRepoFromGitRemoteURL(url: string) {
   return url
     ?.replace(/^git@github.com:|.git$/gims, '')
     ?.replace(/^https:\/\/github.com\/|.git$/gims, '')
